+++ build.gradle
@@ -455,16 +455,18 @@ task testConnected(type: Test) {
     testLogging {
         events = ["passed", "failed", "skipped", "started", "standard_out"]
     }
-    if (isCiServer) {
-        retry {
-          maxRetries = 3
-          maxFailures = 15
+
+    retry {
+        if (isCiServer) {
+            maxRetries = 3
+            maxFailures = 15
         }
-        maxParallelForks = 2
-    } else {
-        maxParallelForks = Runtime.runtime.availableProcessors().intdiv(2) ?: 1
+        else {
+            maxRetries = 2
+        }
     }
 
+    maxParallelForks = 2
     outputs.upToDateWhen { false }
     maxHeapSize = "2g"
 }
@@ -476,15 +478,17 @@ task testIntegration(type: Test) {
     testLogging {
         events = ["passed", "failed", "skipped", "started", "standard_out"]
     }
-    if (isCiServer) {
-        retry {
-          maxRetries = 3
-          maxFailures = 15
+
+    retry {
+        if (isCiServer) {
+            maxRetries = 3
+            maxFailures = 15
         }
-        maxParallelForks = 2
-    } else {
-        maxParallelForks = Runtime.runtime.availableProcessors().intdiv(2) ?: 1
+        else {
+            maxRetries = 2
+        }
     }
+    maxParallelForks = 2
     outputs.upToDateWhen { false }
 }
